Spin out update_containerplat code
