+++ infra/vm/get_containerplat.py
@@ -1,27 +1,12 @@
-from contextlib import contextmanager
 import os
 import subprocess
 import json
-import tempfile
 from typing import Optional
-import zipfile
 import sys
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
 
 from infra.vm.build_hcsshim import build_hcsshim
-
-
-@contextmanager
-def update_package(package_path: str):
-    with tempfile.TemporaryDirectory() as unzipped_dir:
-        with zipfile.ZipFile(package_path, "a") as zip_ref:
-            zip_ref.extractall(unzipped_dir)
-            yield unzipped_dir
-            for root, dirs, files in os.walk(unzipped_dir):
-                for file in files:
-                    file_path = os.path.join(root, file)
-                    zip_ref.write(file_path, os.path.relpath(file_path, unzipped_dir))
 
 
 def get_containerplat(
